+++ src/IO.hh
@@ -38,7 +38,8 @@ public:
     int Read (char  *Buffer, long long  Offset, int  Length);
 
 #if defined(HAVE_READV)
-    virtual ssize_t ReadV (const XrdOucIOVec *readV, size_t n);
+    virtual  int  ReadV (const XrdOucIOVec *readV, int n);
+
 #endif
 
     int Sync() {return 0;}
